user_data/copilot_v2/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
